+++ src/utils/systemUtils.ts
@@ -1,3 +1,4 @@
+import { execSync } from 'child_process';
 import { platform } from 'os';
 
 import { OperatingSystems } from './types';
@@ -15,4 +16,25 @@ export function getOperatingSystem() {
   }
 
   return OPERATING_SYSTEM;
+}
+
+export function getCompilerArchitecture(compiler: string) {
+  const command = `${compiler} -dumpmachine`;
+  let byteArray: Buffer | undefined;
+
+  try {
+    byteArray = execSync(command);
+  } catch (err) {
+    byteArray = Buffer.from('x64', 'utf-8');
+  }
+
+  const str = String.fromCharCode(...byteArray);
+
+  let isArmArchitecture: boolean = false;
+
+  if (str.toLowerCase().includes('arm')) {
+    isArmArchitecture = true;
+  }
+
+  return isArmArchitecture;
 }